refactor classes from step0
