+++ src/main/java/study/firstweek/firststep/Calculator.java
@@ -1,26 +1,24 @@
-package study.first;
-
-import static study.first.StringUtils.isEmpty;
-import static study.first.StringUtils.isEvenInputLength;
-import static study.first.StringUtils.isNotEmpty;
+package study.firstweek.firststep;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 class Calculator {
 
+    private final StringUtils utils = new StringUtils();
+
     int execute(String input) {
-        if (isEmpty(input)) {
+        if (utils.isEmpty(input)) {
             throw new IllegalArgumentException("Not valid input");
         }
 
-        List<String> convertedList = convertEachCharactorToStringList(input);
+        List<String> convertedChars = convertEachCharactorToStringList(input);
 
-        if (isEvenInputLength(convertedList.size())) {
+        if (utils.isEvenInputLength(convertedChars)) {
             throw new IllegalArgumentException("Not valid input");
         }
 
-        return calculate(convertedList);
+        return calculate(convertedChars);
     }
 
     private int calculate(List<String> inputList) {
@@ -34,8 +32,7 @@ class Calculator {
     private List<String> convertEachCharactorToStringList(String input) {
         return input.chars()
             .mapToObj(i -> String.valueOf((char) i))
-            .filter(i -> isNotEmpty(i))
+            .filter(i -> utils.isNotEmpty(i))
             .collect(Collectors.toList());
     }
-
 }
